renamed folder from terms and conditions to legal
