+++ src/js/components/legal/index.js
